app/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
tensorflow==2.10.0
numpy>=1.22,<2.0
pillow>=9.0
python-multipart>=0.0.6